+++ rust/mlt-core/src/encoder/stream/optimizer.rs
@@ -116,9 +116,9 @@ impl DataProfile {
 
         let target = sample_size(values.len());
         let sample = block_sample(values, target);
-
-        let profile = Self::profile::<T>(sample);
-        profile.candidates(T::zero().count_zeros() == 32)
+        // TODO: remove this
+        assert_eq!(size_of::<T>() == 4, T::zero().count_zeros() == 32);
+        Self::profile::<T>(sample).candidates(size_of::<T>() == 4)
     }
 
     /// Return the list of `Encoder` variants worth trying for `u32` data given the